Commit the remaining ArtFlow work and make the project buildable from a fresh clone
Adds the previously-missing Gradle wrapper (and un-ignores its jar so ./gradlew works), project file persistence and canvas rasterization, the JNI brush header, and the adaptive launcher icon moved to mipmap-anydpi-v26 with its drawables. Aligns the brush, layer, selection and docs with the code that actually exists.

+++ app/src/main/java/com/artflow/studio/core/brush/TextureMapper.kt
@@ -167,9 +167,9 @@ class TextureMapper @Inject constructor() {
         
         // Tint the mask with brush color
         val colorFilter = ColorMatrixColorFilter(floatArrayOf(
-            0f, 0f, 0f, 0f, android.graphics.Color.red(color),
-            0f, 0f, 0f, 0f, android.graphics.Color.green(color),
-            0f, 0f, 0f, 0f, android.graphics.Color.blue(color),
+            0f, 0f, 0f, 0f, android.graphics.Color.red(color).toFloat(),
+            0f, 0f, 0f, 0f, android.graphics.Color.green(color).toFloat(),
+            0f, 0f, 0f, 0f, android.graphics.Color.blue(color).toFloat(),
             0f, 0f, 0f, params.opacity, 0f
         ))
